Added AI players to board - need to add movement still. Moved views to their own python files
